Rename lib/lib to lib/util & rearrange files
