+++ lib/mixins/filters.js
@@ -9,7 +9,7 @@
  */
 var _ = require( 'lodash' );
 var extend = require( 'node.extend' );
-var alphaNumericSort = require( '../lib/alphanumeric-sort' );
+var alphaNumericSort = require( '../util/alphanumeric-sort' );
 
 var filterMixins = {};
