+++ lib/shared/wp-request.js
@@ -15,7 +15,7 @@ var _ = require( 'lodash' );
 var extend = require( 'node.extend' );
 
 // TODO: reorganize library so that this has a better home
-var alphaNumericSort = require( '../lib/alphanumeric-sort' );
+var alphaNumericSort = require( '../util/alphanumeric-sort' );
 
 /**
  * WPRequest is the base API request object constructor
