+++ lib/util/build-route-tree.js
