+++ lib/util/generate-path-part-setter.js
