+++ lib/util/named-group-regexp.js
