+++ lib/util/parse-route-string.js
@@ -13,13 +13,13 @@ var _ = require( 'lodash' );
 // 	}) );
 // }
 
-var CollectionRequest = require( './shared/collection-request' );
+var CollectionRequest = require( '../shared/collection-request' );
 var inherit = util.inherits;
 var extend = require( 'node.extend' );
 var generatePathPartSetter = require( './generate-path-part-setter' );
 
 // All valid routes in API v2 beta 11
-var routes = require( './endpoint-response.json' ).routes;
+var routes = require( '../data/endpoint-response.json' ).routes;
 
 var getValues = require( 'lodash' ).values;
